Add support for all supported headers http://docs.oasis-open.org/amqp/core/v1.0/os/amqp-core-messaging-v1.0-os.html#type-header

+++ src/main/java/io/vertx/ext/amqp/impl/AmqpMessageImpl.java
@@ -249,8 +249,8 @@ public class AmqpMessageImpl implements AmqpMessage {
   }
 
   @Override
-  public long deliveryCount() {
-    return message.getDeliveryCount();
+  public int deliveryCount() {
+    return (int) message.getDeliveryCount();
   }
 
   @Override
